refactored txt_metadata test
